break out qualifierMappings, fix bugs
